requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pyTelegramBotAPI==4.12.0
requests==2.31.0
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
python-dotenv==0.21.0